Use framework weight materialization (bf16), add GPU memory tracing
- spike: switch from manual safe_open bf16→f32 to rt._materialize_weight()
  (halves GPU memory: 6.5GB → 3.3GB)
- Add allocation_stats sampling to decode loops in both execution.py and
  spike run.py for memory curve comparison
- Regenerate shaders/tensors for bf16 weight dtype
- Delete unused export_v3.py

+++ src/models/spike_qwen3_asr/shaders/decode_embed_export_embedding_f32.py
@@ -26,7 +26,7 @@ DECODE_EMBED_EXPORT_EMBEDDING_F32 = ShaderVariant(
                 name='weight',
                 io_kind=IOKind.INPUT,
                 role='weight',
-                contract=TensorContract(dtype='float32', shape=('W0', 'W1',)),
+                contract=TensorContract(dtype='bfloat16', shape=('W0', 'W1',)),
             ),
             TensorFieldSpec(
                 name='indices',
@@ -54,8 +54,9 @@ DECODE_EMBED_EXPORT_EMBEDDING_F32 = ShaderVariant(
     execution_requirements=None,
     source="""\
 #version 450
+#extension GL_EXT_bfloat16 : require
 layout(std430) buffer;
-layout(set = 0, binding = 0) buffer restrict readonly WeightBuffer { float weight[]; };
+layout(set = 0, binding = 0) buffer restrict readonly WeightBuffer { bfloat16_t weight[]; };
 layout(set = 0, binding = 1) buffer restrict readonly IndicesBuffer { int indices[]; };
 layout(set = 0, binding = 2) buffer restrict writeonly OutputBuffer { float output_values[]; };
 layout(push_constant) uniform PushConstants { uint num_indices; uint embedding_dim; } pc;
@@ -66,7 +67,7 @@ void main() {
     const uint token_idx = idx / pc.embedding_dim;
     const uint dim_idx = idx - token_idx * pc.embedding_dim;
     const int token_id = indices[token_idx];
-    output_values[idx] = weight[uint(token_id) * pc.embedding_dim + dim_idx];
+    output_values[idx] = float(weight[uint(token_id) * pc.embedding_dim + dim_idx]);
 }
 """,
 )
