+++ src/models/spike_qwen3_asr/tensors/text.py
@@ -35,7 +35,7 @@ def create_embed_tokens(prefix: str, *, bindings: Mapping[str, LogicalTensor] | 
             'p_weight',
             _declare_tensor(
             name="thinker.model.embed_tokens.weight",
-            spec=TensorSpec(dtype='float32', shape=(151936, 1024)),
+            spec=TensorSpec(dtype='bfloat16', shape=(151936, 1024)),
             role=TensorRole.WEIGHT,
             memory=MemoryClass.MODEL_WEIGHT,
             lifetime=TensorLifetime.MODEL,
@@ -165,7 +165,7 @@ def create_text_norm(prefix: str, *, bindings: Mapping[str, LogicalTensor] | Non
             'p_weight',
             _declare_tensor(
             name="thinker.model.norm.weight",
-            spec=TensorSpec(dtype='float32', shape=(1024,)),
+            spec=TensorSpec(dtype='bfloat16', shape=(1024,)),
             role=TensorRole.WEIGHT,
             memory=MemoryClass.MODEL_WEIGHT,
             lifetime=TensorLifetime.MODEL,
@@ -304,7 +304,7 @@ def create_lm_head(prefix: str, *, bindings: Mapping[str, LogicalTensor] | None 
             'p_weight',
             _declare_tensor(
             name="thinker.lm_head.weight",
-            spec=TensorSpec(dtype='float32', shape=(151936, 1024)),
+            spec=TensorSpec(dtype='bfloat16', shape=(151936, 1024)),
             role=TensorRole.WEIGHT,
             memory=MemoryClass.MODEL_WEIGHT,
             lifetime=TensorLifetime.MODEL,
